+++ src/lib.rs
@@ -1,0 +1,13 @@
+
+#[macro_use]
+extern crate log;
+#[macro_use]
+extern crate num_derive;
+mod main;
+mod commands;
+mod config;
+mod db;
+mod errors;
+mod protocol;
+#[cfg(test)]
+mod test_util;
