+++ tests/integration_test.rs
@@ -1,22 +1,24 @@
-/// All tests in this folder assume a server running on localhost:4242
+#[macro_use]
+extern crate lazy_static;
 use futures::SinkExt;
 use serde::{Deserialize, Serialize};
 use tokio::net::TcpStream;
 use tokio::stream::StreamExt;
 use tokio_util::codec::{Framed, LengthDelimitedCodec};
-
+use std::panic;
 use bytes::Bytes;
+use std::process::Child;
 
 mod common;
 
-static LOCAL_REMITS: &str = "localhost:4243";
+static LOCAL_REMITS: &str = "0.0.0.0:4243";
 static OK_RESP: &[u8] = &[0x62, 0x6F, 0x6B];
+
 
 #[tokio::test]
 async fn integration_tests() {
-    let server = common::start_server();
+    common::start_server();
     let framer = &mut (connect_to_remits().await);
-
     println!("test: should be able to add a log");
     let (kind, code, payload) = send_req(framer, new_log_add_req("test")).await;
     assert_eq!(kind, 0x01);
@@ -84,7 +86,6 @@ async fn integration_tests() {
     assert_eq!(code, 0x00);
     let out: Vec<String> = serde_cbor::from_slice(&payload[4..]).unwrap();
     assert_eq!(out, vec!("test"));
-    common::teardown(&server);
 }
 
 fn new_log_add_req(name: &str) -> Vec<u8> {
